--- src/encrypted_comm/__init__.py
@@ -1,2 +1,0 @@
-"""Encrypted Communication Module
-"""
